Created map functionallity, did not intgrate this into game yet
Player can move around a toroid map, but can not do so in game. Next task will be creating move functionallity after/before fights. Adding terrain such as rivers and mountains is also needed. AI will need to navigate map, thus it is it's own file so that the AI can call it in it's AI tree.